v3.0: Change completion API to work with any learning context
This changes the python API of edx-completion to use the more generic concept of "learning context" instead of courses. This allows the app to track the completion of XBlocks that are in content libraries or other future learning contexts and not just courses.

I decided to do a clean and backwards-incompatible change to the entire python API rather than trying to preserve the existing API in a backwards-compatible way because this python API is relatively new, is only really consumed by edx-platform which I'll update at the same time, and because having a mix of "course_key" and "context_key" throughout the code was getting a bit messy.

While I was at it, I simplified the API, removing the need to specify a course_key and a block_key, as the course_key/context_key can be trivially derived from the block_key and the only exception is for old_mongo which is deprecated and can still be supported with these changes as long as the client sends the right block_key format.

This includes a data migration but it produces no actual database changes.

I have kept the REST API unchanged since changes to REST APIs are more problematic, especially if used by deployed mobile apps.