fix: Pylint failures in version upgrade 2.13.8 -> 2.15.10
This commit adds some code to ignore the Pylint checker unnecessary-dunder-call in the CompletionWaffleTestMixin and CompletionSetUpMixin classes, which both provide a method to override a waffle switch. Changing the code to satisfy this checker would involve a rewrite of the CompletionWaffleTestMixin and the CompletionSetUpMixin classes, which are used in integration testing in the completion and compeltion_integration repositories. Both classes are intentionally written this way.